New servo, moved pin inits

+++ src/Servo.h
@@ -16,8 +16,8 @@ private:
 	DigitalIoPin *pin;
 	int portNum;
 	int pinNum;
-	const int penUp_cycle_length = 1300;
-	const int penDown_cycle_length = 1450;
+	const uint16_t penUp_cycle_length = 1300;
+	const uint16_t penDown_cycle_length = 1450;
 	void init();
 public:
 	Servo(int portNum, int pinNum);
